Issue: bold.public.search function query terms with no records on BOLD
Fixed the issue in the bold.public.search function query search. Earlier version used an 'all or nothing' approach wherein if a term that might not have had any data was present in a vector of query terms, whole download used to be aborted. Now the search result gives the list of ids for all the available search queries and prints the terms for which data could not be retrieved